requirements.txt
streamlit
pandas
matplotlib
miepython
netCDF4
scipy
seaborn
xarray
pytest
tqdm
pyyaml
plotly
.
